bug(tests): Fix local test scripts (#316)
The test scripts in scripts didn't work due to missing environment variables needed by the tests.

The tests also relied on the working directory being set such that test artifacts in source could be read. This has been fixed with include_bytes! and include_str! isntead, and the the working directory for the test is no longer modified.